port r931457 to LANG 2.x branch - Bug in copy and paste, now tests MutableFloat not MutableDouble for String constructor
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/lang/branches/LANG_2_X@1057020 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/lang/mutable/MutableFloatTest.java
@@ -39,7 +39,7 @@ public class MutableFloatTest extends TestCase {
         assertEquals(2f, new MutableFloat(new Float(2f)).floatValue(), 0.0001f);
         assertEquals(3f, new MutableFloat(new MutableFloat(3f)).floatValue(), 0.0001f);
 
-        assertEquals(2f, new MutableDouble("2.0").floatValue(), 0.0001f);
+        assertEquals(2f, new MutableFloat("2.0").floatValue(), 0.0001f);
 
         try {
             new MutableFloat((Number)null);
